+++ docs/source/conf.py
@@ -50,9 +50,9 @@ copyright = '2019, Tyler J Nigon'
 author = 'Tyler J Nigon'
 
 # The short X.Y version
-version = '0.1.4'
+version = '0.2.0'
 # The full version, including alpha/beta/rc tags
-release = '0.1.4'
+release = '0.2.0'
 
 # -- General configuration ---------------------------------------------------
 # If your documentation needs a minimal Sphinx version, state it here.